Create test for the general shopping list

+++ db/migrate/20231008170923_create_foods.rb
@@ -5,8 +5,8 @@ class CreateFoods < ActiveRecord::Migration[7.0]
     create_table :foods do |t|
       t.string :name
       t.string :measurement_unit
-      t.integer :price
-      t.integer :quantity
+      t.float :price
+      t.float :quantity
       t.references :user, null: false, foreign_key: true
 
       t.timestamps
